fix unit-test
